Clean up field element
