+++ Tests/Tests/FORMFieldElementTests.m
@@ -15,8 +15,6 @@
                                                                                     @"hidden": @NO,
                                                                                     @"size": @{@"width": @30,
                                                                                                @"height": @1},
-                                                                                    @"position": @1,
-                                                                                    @"value": @"John Malkobitch",
                                                                                     @"minimum_date":@"2000-01-01T00:00:00.002Z",
                                                                                     @"maximum_date":@"2015-01-01T00:00:00.002Z",
                                                                                     @"validations": @{@"required": @YES,
@@ -28,8 +26,6 @@
     XCTAssertEqualObjects(fieldElement.info, @"Display name");
     XCTAssertFalse(fieldElement.hidden);
     XCTAssertTrue(CGSizeEqualToSize(fieldElement.size, CGSizeMake(30, 1)));
-    XCTAssertEqualObjects(fieldElement.position, @1);
-    XCTAssertEqualObjects(fieldElement.value, @"John Malkobitch");
     XCTAssertEqualObjects([fieldElement.minimumDate hyp_dateString], @"2000-01-01");
     XCTAssertEqualObjects([fieldElement.maximumDate hyp_dateString], @"2015-01-01");
     XCTAssertNotNil(fieldElement.validation);
